Respect distance units when ordering by distance
While the unit shouldn't change the result, if you order and select with
a distinct call (ie. `Model.distinct.selecting_distance_from(...).order_by_distance(...)`),
Postgres requires all ordering values to also be selected.

Previously, in such a case, we'd be selecting by miles when that was the
configured unit, but then ordering by meters, which Postres didn't like.

Now, we'll both select miles and order by miles when configured as such.